Add a basic ExpressJS app
ExpressJS is used for doing some backend calls that'll be needed by the
homepage that wouldn't necessarily be "safe" to do on the client side
(because it exposes some API keys that visitors of the homepage
shouldn't see).

One of these backend calls is a call to Steam's Web API to see what I'm
currently playing. :sweat_smile: